feat(task-col): :sparkles: Display of columns and tasks related to a certain project
#33, #34, #36
