+++ project/components/ui/kaban_ui/kanban-board.tsx
@@ -126,24 +126,13 @@ const initialColumns = [
 export function KanbanBoard({ projectId }: { projectId: string }) {
   const{columns,isLoading,error}=useColumns(projectId)
   //const [columns, setColumns] = useState(initialColumns)
-  console.log("Columns found",columns)
+
 
   if (isLoading) return <p>Loading...</p>;
   if (error) return <p>Failed to load columns {error.message}</p>;
   if (!columns) return <p>Failed to load columns</p>;
 
-  const getPriorityColor = (priority: string) => {
-    switch (priority) {
-      case "high":
-        return "bg-red-100 text-red-700 dark:bg-red-900 dark:text-red-300"
-      case "medium":
-        return "bg-yellow-100 text-yellow-700 dark:bg-yellow-900 dark:text-yellow-300"
-      case "low":
-        return "bg-green-100 text-green-700 dark:bg-green-900 dark:text-green-300"
-      default:
-        return "bg-gray-100 text-gray-700 dark:bg-gray-800 dark:text-gray-300"
-    }
-  }
+
 
   return (
             <div className="bg-white dark:bg-outer_space-500 rounded-lg border border-french_gray-300 dark:border-payne's_gray-400 p-6">
